requirements.txt
aiohttp==3.9.1
attrs==23.1.0
blinker==1.7.0
branca==0.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
Flask==3.0.0
idna==3.6
ipinfo==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
urllib3==2.1.0
Werkzeug==3.0.1